% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOJfame.R
\name{get_bojfame}
\alias{get_bojfame}
\title{Download and parse data from BOJ Time-Series Data Search portal}
\usage{
get_bojfame(
  series_codes,
  start_year = NULL,
  end_year = NULL,
  base_url = NULL,
  auto_pivot = TRUE,
  ...
)
}
\arguments{
\item{series_codes}{Character. A vector of series codes to retrieve. All
codes must correspond to data of the same frequency.}

\item{start_year}{Numeric. Start year in format: YYYY (optional).}

\item{end_year}{Numeric. End year in format: YYYY (optional).}

\item{base_url}{Character. URL for generating the BOJ data download page
(optional).}

\item{auto_pivot}{Logical. Controls whether source data set is converted to
long format. Set this to \code{FALSE} to disable conversion (default: TRUE).}

\item{...}{Arguments passed to \code{download.file()} (e.g.
\code{quiet = TRUE}).}
}
\value{
A list containing the time series data (\code{$data}), corresponding
meta data (\code{$meta}), the URL of the download page (\code{$page_url}),
and the URL of the CSV file ((\code{$item_url})).
}
\description{
Download and parse data from BOJ Time-Series Data Search portal
}
\details{
If \code{start_year} and \code{end_year} are omitted, the last five
years of data (or the last year for daily series) will be downloaded.
}
\examples{
\donttest{
df <- get_bojfame(series_codes = c("FM01'STRDCLUCON", "IR01'MADR1Z@D"))
}
}
