% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BOJfame.R
\name{pivot_longer_bojfame}
\alias{pivot_longer_bojfame}
\title{Convert a BOJ FAME data set to long format}
\usage{
pivot_longer_bojfame(tbl)
}
\arguments{
\item{tbl}{Tibble. A tibble data frame holding a BOJ FAME data set (usually
obtained via \code{get_bojfame(series_codes, auto_pivot = FALSE)}).}
}
\value{
A tibble data frame.
}
\description{
Convert a BOJ FAME data set to long format
}
\examples{
\donttest{
bojdata <- get_bojfame(series_codes = c("FM01'STRDCLUCON", "IR01'MADR1Z@D"),
                       auto_pivot = FALSE)
bojdata$data <- pivot_longer_bojfame(bojdata$data)
}
}
